+++ src/core/gui.model.ts
@@ -1,8 +1,10 @@
 import { FilesManager } from './files';
 import { SchematicsManager } from './schematics';
+import { Subject } from 'rxjs/Subject';
 import { defaultConfiguration } from './models';
 
 export interface GUI {
+  action: Subject<any>;
   cliConfig;
   cliCollection: string;
   converter;