+++ src/local.ts
@@ -1,4 +1,5 @@
 import * as path from 'path';
+import * as vscode from 'vscode';
 
 import { AngularGUI, defaultConfiguration as config } from './core';
 
@@ -8,8 +9,8 @@ import { AngularGUI, defaultConfiguration as config } from './core';
  * for me it's "webapps", change this to match your setup
  */
 config[ 'workspaceRoot' ] = path.resolve(__dirname, '..', '..', 'webapps');
-config[ 'extensionRoot' ] = path.resolve(__dirname, '..');
+config[ 'extensionRoot' ] = path.resolve(__dirname, '..', '.angular-gui');
 config[ 'local' ] = true;
 
 const gui = new AngularGUI(config, console.log);
-gui.start(o => o);
+gui.start(o => o).subscribe(console.log);